==> 03-cpt.Rmd <==
# Réponse à la Camptothecine {#cpt}

```{r library2, warning = FALSE, message = FALSE}
library(tidyr)
library(stringr)
library(FactoMineR)
library(ggplot2)
library(ggsci)
library(hrbrthemes)
```

```{r recuperation2, echo = FALSE}
d_merge = readRDS(file = "data/d_merge.rds")
```

## Réponse des mutants à CPT

```{r histo, message = FALSE}
cpt = data.frame(ORF = d_merge$ORF, Reponse_cpt = d_merge$camptothecin_500um)
ggplot(cpt) +
  aes(x = Reponse_cpt) +
  theme_ipsum()+
  #geom_histogram(aes(y = after_stat(density)), color = "grey", fill = "#4DBBD5FF") +
  labs(title = "Réponses à la Camptothécine par stress", x = "Perte de fitness (log)", y = "Densité", subtitle = "Parmis les 448 stress distincts")+
  geom_density(color = "grey", fill = "#E64B35FF", alpha = 0.4)
```

```{r}
cpt$ORF[cpt$Reponse_cpt > 2.5] %>% length()
cpt$ORF[cpt$Reponse_cpt > 2] %>% length()
cpt$ORF[cpt$Reponse_cpt > 1.5] %>% length()
```
```{r}
cpt$ORF[cpt$Reponse_cpt < -1] %>% length()
cpt$ORF[cpt$Reponse_cpt < -.8] %>% length()
cpt$ORF[cpt$Reponse_cpt < -.7] %>% length()
cpt$ORF[cpt$Reponse_cpt < -.6] %>% length()
cpt$ORF[cpt$Reponse_cpt < -.5] %>% length()
```

## GO enrichissement

En utilisant https://yeastgenome.org/goTermFinder, on regarde si le groupe des gènes dont la mutation entraine une plus grande sensibilité à cpt est enrichis par rapport au génome ebntier dans certaines fonctions moléculaire, localisations cellulaires ou processus biologiques :

### Mutants les plus resistants à CPT

```{r}
# les 43 plus resistant à cpt
write.table(data.frame(ORF = d_merge$Gene, Reponse_cpt = d_merge$camptothecin_500um)$ORF[cpt$Reponse_cpt > 1.5], file = "data/gene_15.txt", row.names = FALSE, quote = FALSE)
```

rien de ouf, marche pas trop

```{r, eval = FALSE, echo = FALSE}
# les 61 plus sensible à cpt
write.table(data.frame(ORF = d_merge$Gene, Reponse_cpt = d_merge$camptothecin_500um)$ORF[cpt$Reponse_cpt < -.5], file = "data/gene_-5.txt", row.names = FALSE, quote = FALSE)
```

```{r, eval = FALSE, echo = FALSE}
tmp = read.table("data/go/61_process.txt", sep = "\t", header = TRUE)
tmp = data.frame(Term = tmp$TERM, p_value = tmp$CORRECTED_PVALUE, FDR = tmp$FDR_RATE, Gene = tmp$ANNOTATED_GENES)

knitr::kable(tmp, caption = "Biological Process")

tmp = read.table("data/go/61_function.txt", sep = "\t", header = TRUE)
tmp = data.frame(Term = tmp$TERM, p_value = tmp$CORRECTED_PVALUE, FDR = tmp$FDR_RATE, Gene = tmp$ANNOTATED_GENES)

knitr::kable(tmp, caption = "Molecular Function")

tmp = read.table("data/go/61_component.txt", sep = "\t", header = TRUE)[c(1:10),]
tmp = data.frame(Term = tmp$TERM, p_value = tmp$CORRECTED_PVALUE, FDR = tmp$FDR_RATE, Gene = tmp$ANNOTATED_GENES)

knitr::kable(tmp, caption = "Cellular Component")
```

### Les 24 plus sensibles

```{r}
# les 24 plus senssible à cpt
write.table(data.frame(ORF = d_merge$Gene, Reponse_cpt = d_merge$camptothecin_500um)$ORF[cpt$Reponse_cpt < -.6], file = "data/gene_-6.txt", row.names = FALSE, quote = FALSE)
```

```{r}
tmp = read.table("data/go/24_process.txt", sep = "\t", header = TRUE)
tmp = data.frame(Term = tmp$TERM, p_value = tmp$CORRECTED_PVALUE, FDR = tmp$FDR_RATE, Gene = tmp$ANNOTATED_GENES)

knitr::kable(tmp, caption = "Biological Process")

tmp = read.table("data/go/24_function.txt", sep = "\t", header = TRUE)
tmp = data.frame(Term = tmp$TERM, p_value = tmp$CORRECTED_PVALUE, FDR = tmp$FDR_RATE, Gene = tmp$ANNOTATED_GENES)

knitr::kable(tmp, caption = "Molecular Function")

tmp = read.table("data/go/24_component.txt", sep = "\t", header = TRUE)
tmp = data.frame(Term = tmp$TERM, p_value = tmp$CORRECTED_PVALUE, FDR = tmp$FDR_RATE, Gene = tmp$ANNOTATED_GENES)

knitr::kable(tmp, caption = "Cellular Component")
```

les mutant "les plus senssible" à cpt sont en réalité des mutants qui ont des fitness defect significatif dans pratiquement toutes les conditions. Ces mutant sont probablement les plus senssible à tout type de stress/drogue car leurs mutations sont graves.

Exemple de RPL30 :

```{r density_rpl30}
RPL30 = d_merge[d_merge$Gene == "RPL30",][-c(1,2)] %>% unlist()

ggplot(data.frame(Fitness_d = RPL30)) +
  aes(x = Fitness_d)+
  geom_density(color = "#4DBBD5FF", fill = "#4DBBD5FF", alpha = 0.4)+
  geom_vline(xintercept = RPL30 %>% mean(), color = "#E64B35FF", linetype = "dashed", size = 1) + #moyenne
  geom_vline(xintercept = d_merge[d_merge$Gene == "RPL30",]$camptothecin_500um, color = "#00A087FF", size = 1) + #CPT response
  labs(title = "Réponse aux stress du mutant RPL30", x = "Fitness Defect (FD)", y = "Densité", subtitle = "Moyenne des réponses (rouge) et réponse à CPT (vert)")+
  theme_ipsum()
```


## Nouvelle approche (pénalisée)

Pour être considéré comme sensible à CPT, un gène doit être plus senssible à CPT qu'aux autres stress :

```{r}
gene = d_merge[d_merge$camptothecin_500um < 0,]$Gene
cpt_d_merge = d_merge[d_merge$camptothecin_500um < 0,][-c(1,2)]
cpt_d_merge %>% dim()

# correction par rapport à la médianne 
tmp_med = cpt_d_merge %>% apply(1, function(x){
  x - median(x)
}) %>% as.data.frame() %>% t() %>% as.data.frame()

tmp_med$camptothecin_500um %>% summary()

```
Marche pas du tout à cause du bazard négatif positif etc. Pas le temps de ce pencher dessus pour de vrai.

Autre approche : prendre pour chaque mutant les 5% stress qui le mette le plus en pls. On part en fait de l'hypothèse que comme CPT est très toxique (introduction), elle est parmis les drogues les plus efficace pour eliminer les souches mutantes dans des voies liées à son mode d'action.

```{r}
# les plus sensible
classement = NULL
nb = 25
cutoff = -0.4
for (i in c(1:(gene %>% length()))) {
  tmp = data.frame(stress = colnames(cpt_d_merge), fd = cpt_d_merge[i,] %>% unlist())
  tmp = tmp[order(tmp$fd),]
  if (classement %>% is.null()){
    classement = data.frame(gene = rep(gene[i], nb), stress = tmp$stress[1:nb], fd = tmp$fd[1:nb])
  }
  else {
    classement = classement %>% rbind(data.frame(gene = rep(gene[i], nb), stress = tmp$stress[1:nb], fd = tmp$fd[1:nb]))
  }
}

selection = classement$gene[classement$stress == "camptothecin_500um" & classement$fd < cutoff]
write.table(selection, file = "data/top_-genes.txt", quote = FALSE, row.names = FALSE)


# Les moins sensibles
gene2 = d_merge[d_merge$camptothecin_500um > 0,]$Gene
cpt_d_merge2 = d_merge[d_merge$camptothecin_500um > 0,][-c(1,2)]
classement = NULL
nb = 15
cutoff = 1
for (i in c(1:(gene2 %>% length()))) {
  tmp = data.frame(stress = colnames(cpt_d_merge2), fd = cpt_d_merge2[i,] %>% unlist())
  tmp = tmp[order(-tmp$fd),]
  if (classement %>% is.null()){
    classement = data.frame(gene = rep(gene2[i], nb), stress = tmp$stress[1:nb], fd = tmp$fd[1:nb])
  }
  else {
    classement = classement %>% rbind(data.frame(gene = rep(gene2[i], nb), stress = tmp$stress[1:nb], fd = tmp$fd[1:nb]))
  }
}

selection = classement$gene[classement$stress == "camptothecin_500um" & classement$fd > cutoff]
write.table(selection, file = "data/top_+genes.txt", quote = FALSE, row.names = FALSE)
```

### GO enrichissement

Ya rien qui marche bon sang

### Preuve que la première approche est fausse

En selectionnant les mutants dont la medianne de fitness defect est la plus basse, on obtient le même enrichissement :

```{r}
med = data.frame(medianne = d_merge[-c(1,2)] %>% apply(1, function(x){median(x)}), gene = d_merge$Gene)
med = med[order(med$medianne),]

week_mut = med$gene[1:40]
write.table(week_mut, file = "data/week_mut.txt", quote = FALSE, row.names = FALSE)

tmp = read.table("data/go/preuve_function.txt", sep = "\t", header = TRUE)
tmp = data.frame(Term = tmp$TERM, p_value = tmp$CORRECTED_PVALUE, FDR = tmp$FDR_RATE, Gene = tmp$ANNOTATED_GENES)

knitr::kable(tmp, caption = "Molecular Function")
```

## Encore une autre approche

Répartition de la médianne des fitness defect de chaque mutant

```{r density_med}
med = d_merge[-c(1,2)] %>% apply(MARGIN = 1, function(x){
  median(x)
})

ggplot(data.frame(medianne = med)) +
  aes(x = medianne)+
  geom_density()
```

```{r second filtre}
seuil_min = -0.1
seuil_max = 0.2
d_merge_filtre = d_merge[med > seuil_min & med < seuil_max,]

cpt_filtre = data.frame(Gene = d_merge_filtre$Gene, Reponse_cpt = d_merge_filtre$camptothecin_500um)
nb_gene = 50
top_x = cpt_filtre[order(cpt_filtre$Reponse_cpt),]$Gene[c(1:nb_gene)]
write.table(top_x, file = "data/top_x_gene.txt", quote = FALSE, row.names = FALSE)
```

Rien d'enrichis non plus... Sauf si on prend un seuil baas plus petit que -0.3 auquel cas on retrouve ce foutu ribosome
