==> 05-classif_suite.Rmd <==
# Classification des stress (suite) {#classif2}

```{r library4, warning = FALSE, message = FALSE}
library(tidyr)
library(stringr)
library(FactoMineR)
library(mclust)
library(umap)
library(ggplot2)
library(ggrepel)
library(RColorBrewer)
library(hrbrthemes)
library(viridis)
library(ggsci)
```

```{r recuperation4, echo = FALSE}
d_merge = readRDS(file = "data/d_merge.rds")
tab = read.csv2("data/All_stress_annotated.csv", sep = "\t", header = TRUE)
y = data.frame(type1 = tab$Type %>% unique(), type2 = c("Cytotoxique", "Divers", "Divers", "Divers", 
                                                        "Divers", "Antifongique",
                                                        "Cytotoxique", "Divers", "Antibactériens / Antiviral", "Divers", "Inhibiteur de kinase",
                                                        "Divers", "Divers", "Inihbiteur de la synthèse protéique", 
                                                        "Inhibiteur de la synthèse d'acide nucléique",
                                                        "Divers", "Divers", "Cytotoxique", "Divers", "Cytotoxique", "Divers",
                                                        "Divers", "Divers", "Antibactériens / Antiviral", "Inhibiteur de phosphatase", 
                                                        "Inhibiteur d'enzyme", "Divers","Divers" , "Divers", "Divers", "Divers",
                                                        "Divers", "Divers", "Divers", "Divers", "Divers", "Divers"))
tab[["Type_merge"]] = tab$Type %>% sapply(function(s){
  y$type2[y$type1 == s]
}) %>% unlist()
gg = readRDS("data/classif_grosse.rds")
```

## Caractérisation des 18 groupes de stress

```{r}
for (i in seq(1,18)) {
  cat("\n")
  print(paste0("STRESS DE LA CLASSE ", i, " :"))
  print(gg[gg$classe == i,]$name)
}
```

## Enrichissement de classe 

Es ce que le groupe de stress contenant la camptothecine est enrichit en un ou plusieurs types de stress ?

```{r}
gg[gg$name == "camptothecin_500um",]

stress_campto <- gg[gg$classe == gg[gg$name == "camptothecin_500um",]$classe,]$name
saveRDS(stress_campto, file = "data/stress_campto.rds")
```

```{r plota}
gg_campto = data.frame(x = stress_campto, Type = stress_campto %>% sapply(function(y){
  tab$Type[y == tab$x]
}))

full_count = data.frame(Type = tab$Type %>% unique(), Count = tab$Type %>% unique() %>% sapply(function(y){
  tab$x[tab$Type == y] %>% length()
}), groupe = rep("ALL", tab$Type %>% unique() %>% length()))
campto_count = data.frame(Type = tab$Type %>% unique(), Count = tab$Type%>% unique()  %>% sapply(function(y){
  gg_campto$x[gg_campto$Type == y] %>% length()
}), groupe = rep("CPT", tab$Type %>% unique() %>% length()))

gg_compar = full_count %>% rbind(campto_count)

ggplot(gg_compar)+
  aes(x = groupe, fill = Type, y = Count) +
  geom_col(position = "fill")
```

```{r plotb, warning=FALSE}
gg_campto = data.frame(x = stress_campto, Type = stress_campto %>% sapply(function(y){
  tab$Type_merge[y == tab$x]
}))

full_count = data.frame(Type = tab$Type_merge %>% unique(), Count = tab$Type_merge %>% unique() %>% sapply(function(y){
  tab$x[tab$Type_merge == y] %>% length()
}), groupe = rep("ALL", tab$Type_merge %>% unique() %>% length()))
campto_count = data.frame(Type = tab$Type_merge %>% unique(), Count = tab$Type_merge %>% unique()  %>% sapply(function(y){
  gg_campto$x[gg_campto$Type == y] %>% length()
}), groupe = rep("CPT", tab$Type_merge %>% unique() %>% length()))

gg_compar = full_count %>% rbind(campto_count)

ggplot(gg_compar)+
  aes(x = groupe, fill = Type, y = Count) +
  #scale_fill_ipsum() +
  #scale_fill_brewer(palette = "Pastel1")+
  #scale_fill_viridis_d(option = "A")+
  scale_fill_npg()+
  geom_col(position = "fill") +
  labs(title = "Types de stress du groupe Camptothécine", x = "Groupe", subtitle = "Comparaison avec les proportions initiales")+
  theme_ipsum()
```

Mesure de l'enrichissement dans chacune des catégories :

On réalise un test de fisher exact pour les 16 catégories (loi hypergéométrique)

```{r testf}

alpha = 0.05
tests = NULL
for (categorie in full_count$Type) {
  m = full_count$Count[full_count$Type == categorie] # boules blanches totales
  x = c(campto_count$Count[campto_count$Type == categorie]:m) # boules blanches tirées 
  n = full_count$Count[full_count$Type != categorie] %>% sum() # boules totales
  k = campto_count$Count %>% sum() # nb tirage
  mat = matrix(c(campto_count$Count[campto_count$Type == categorie], 
                 k-campto_count$Count[campto_count$Type == categorie], 
                 m-campto_count$Count[campto_count$Type == categorie], 
                 n-(k-campto_count$Count[campto_count$Type == categorie])), 
               nrow = 2)
  
  tests = c(tests, dhyper(x, m, n, k, log = FALSE) %>% sum())
  #équivaut à fisher.test(mat)$p.value
}
#Correction de Bonferonni
#tests = tests / (full_count$Type %>% length())

tf = data.frame(categories = full_count$Type, p_val = tests)
tf[tf$p_val < alpha,]

```
Ce groupe est enrichis en composants **cyctotoxiques** et **inhibiteurs de la synthèse d'acides nucléiques**. Ces deux points coincident avec les trouvailles de la GO enrichissement.
